smooth correction

+++ Fractal.hpp
@@ -14,6 +14,6 @@ private:
     int maxIterations;
     std::vector<unsigned char> image; // RGB pixels
 
-    int computeIterations(double x0, double y0);
-    void mapColor(int iterations, int maxIterations, unsigned char& r, unsigned char& g, unsigned char& b);
+    double computeIterations(double x0, double y0);
+    void mapColor(double iterations, int maxIterations, unsigned char& r, unsigned char& g, unsigned char& b);
 };
